- move hpp files to include/opencv2/ccalib - add multiple camera calibration tutorial - code clean
