+++ modules/ccalib/sample/omniStereoCalibration.cpp
@@ -1,4 +1,4 @@
-#include "opencv2/omnidir.hpp"
+#include "opencv2/ccalib/omnidir.hpp"
 #include "opencv2/core.hpp"
 #include "opencv2/imgproc.hpp"
 #include "opencv2/highgui.hpp"
