+++ modules/ccalib/sample/randomPatternCalibration.cpp
@@ -1,4 +1,4 @@
-#include "opencv2/randomPatten.hpp";
+#include "opencv2/ccalib/randomPatten.hpp";
 #include "opencv2/highgui.hpp"
 #include "opencv2/imgproc.hpp"
 #include "opencv2/calib3d.hpp"
